+++ but-what-about/CHANGELOG.md
@@ -2,8 +2,24 @@
 
 #### New Features
 
-* Added stairs, slabs, and walls for Calcite, Cracked Deepslate Bricks, Cracked Deepslate Tiles,
-  Cracked Nether Bricks, Cracked Polished Blackstone Bricks, Cracked Stone Bricks, End Stone,
-  Netherrack, and Smooth Basalt.
+* Added stairs, slabs, and walls for
+  * Calcite
+  * Chiseled Copper (Exposed, Weathered, Oxidized, and waxed of each)
+  * Chiseled Deepslate
+  * Chiseled Nether Bricks
+  * Chiseled Polished Blackstone
+  * Chiseled Sandstone
+  * Chiseled Red Sandstone
+  * Chiseled Resin Bricks
+  * Chiseled Stone Bricks
+  * Chiseled Sulfur
+  * Cracked Deepslate Bricks
+  * Cracked Deepslate Tiles
+  * Cracked Nether Bricks
+  * Cracked Polished Blackstone Bricks
+  * Cracked Stone Bricks
+  * End Stone
+  * Netherrack
+  * Smooth Basalt
 
 #### Changes